+++ test/integration/gardenlet/seed/lease/lease_test.go
@@ -18,7 +18,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
-	leasereconciler "github.com/gardener/gardener/pkg/gardenlet/controller/seed/lease"
+	leasecontroller "github.com/gardener/gardener/pkg/gardenlet/controller/lease"
 	"github.com/gardener/gardener/pkg/utils/test"
 	. "github.com/gardener/gardener/pkg/utils/test/matchers"
 )
@@ -69,7 +69,7 @@ var _ = Describe("Seed lease controller tests", func() {
 
 	It("should not update the Lease object and set the internal health status to false", func() {
 		var fakeError error
-		DeferCleanup(test.WithVar(&leasereconciler.CheckSeedConnection, func(context.Context, rest.Interface) error { return fakeError }))
+		DeferCleanup(test.WithVar(&leasecontroller.CheckSeedConnection, func(context.Context, rest.Interface) error { return fakeError }))
 
 		By("Ensure Lease got maintained")
 		Eventually(func(g Gomega) {
